requirements.txt
agentic-ai==0.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
groq==0.18.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.5
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kaleido==0.2.1
langchain==0.2.17
langchain-core==0.2.43
langchain-ollama==0.1.3
langchain-text-splitters==0.2.4
langsmith==0.1.147
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.15.1
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
ollama==0.4.2
openai==1.61.1
orjson==3.10.12
outcome==1.3.0.post0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil_resolve_name==1.3.10
playwright==1.48.0
plotly==5.24.1
propcache==0.2.0
protobuf==5.29.0
pyarrow==17.0.0
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.20.1
safetensors==0.4.5
selenium==4.27.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
streamlit==1.40.1
streamlit-chat==0.1.1
sympy==1.13.3
ta==0.11.0
tenacity==8.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
trio==0.27.0
trio-websocket==0.11.1
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.15.2
zipp==3.20.2
